requirements.txt
Flask==2.2.2
psycopg2-binary==2.9.5
pytest==7.2.1
pytest-cov==4.0.0